Consolidate and simplify atomic operations
A template ApplyAMO(flags, target, value) applies the atomic operation in
flags to target with value. It works with both 32-bit and 64-bit values.
